+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Search metadata for category events.
+
 ## [2.125.0] - 2022-05-17
 ### Added
 - Support for different settings by binding.